Full Project
After taking a look at different languages that compile to JS to
write this in, I settled on TypeScript instead of ReasonML.

- Rewrote in TypeScript
- Added rollup and some plugins
- Added eslint and @coalpha/eslint-config
- Moved the index.html demo stuff to example/

I still need to finish the getters and setters. I think that the
private `#` syntax might be a bit new to use right now.
It might not be horrible to just expose the canvas and stuff.
I'll just declare that you get UB if you decide to mess with 'em.

Once all of that is done and I'm happy with the API, I'll profile
it and rewrite hot paths in AssemblyScript.