FPROverflow
- Fixed a bug where all the public variables weren't initialized
before callbacks were called. (bug was in #resize)
- Added overflow mode.
- Haven't fixed that bug where the canvas doesn't resize correctly
after zooming in one direction twice and then reversing.

That overflow thing might be a good idea.
I'm thinking that I could make a custom one where it adds the
canvas directly to the <body> and is the size of the entire window.
That way it never has to change unless the Window is zoomed.

I'll probably call it fullscreen-canvas or something.
It'd be better if I could just query the size of the entire screen
in DevicePixels but that seems to have different implementations
across browsers... lookin' at you, chrome.